+++ src/components/skill/Card.tsx
