+++ src/components/values/Card.tsx
